sqlite_client/requirements.txt
mcp>=1.3.0
anthropic>=0.49.0
loguru>=0.7.3
python-dotenv>=1.0.1
pytest>=7.4.0
pytest-asyncio>=0.21.1 